refactor: split the slides deck into per-slide components
src/routes/slides/+page.svelte (1991 lines) becomes:
- $lib/slides/: SlideCover/Problem/Vision/Audience/Journey/How/Strengths/
  Close, Logomark, Pipeline (pnode+serpentine snippets), typed data.ts, and
  slides.css (shared deck styles, namespaced under .deck with deck- prefixed
  keyframes so nothing leaks across SPA navigations)
- $lib/utils/tilt.ts: use:tilt action replacing the copy-pasted pointer
  handlers; reduced-motion aware
- page drops to 353 lines (head, deck state, keyboard nav, controls)
- dead .wave/.modules styles removed — svelte-check is now warning-free

No visual change (verified against desktop + mobile screenshots of all 8
slides on the built preview).
