+++ mono/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "mono",
 	"private": true,
-	"version": "2.11.5",
+	"version": "2.11.6",
 	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
