+++ mono/src/lib/slides/SlideClose.svelte
@@ -1,0 +1,18 @@
+<script lang="ts">
+  import { ArrowUpRight } from "lucide-svelte";
+  import Logomark from "./Logomark.svelte";
+</script>
+
+<div class="slide-body center cover">
+  <span class="logomark big r" style="--i:0" aria-hidden="true">
+    <Logomark size={26} />
+  </span>
+  <h2 class="title r" style="--i:1">Exams, end to end.</h2>
+  <p class="message center-msg r" style="--i:3">
+    One workspace for the whole exam lifecycle — quiet, private, complete.
+  </p>
+  <a class="cta r" style="--i:4" href="/">
+    Open TAO <ArrowUpRight size={17} strokeWidth={2} />
+  </a>
+  <p class="hint r" style="--i:5">tao.lv0.eu · built for the SPF Finances test team</p>
+</div>
