+++ mono/src/lib/slides/SlideCover.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  import Logomark from "./Logomark.svelte";
+</script>
+
+<div class="slide-body center cover">
+  <div class="brand r" style="--i:0">
+    <span class="logomark" aria-hidden="true"><Logomark size={30} /></span>
+    <h1 class="wordmark" data-text="TAO">TAO</h1>
+  </div>
+  <p class="lead r" style="--i:1">Exams, end to end.</p>
+  <p class="sub r" style="--i:3">A unified exam toolkit</p>
+  <p class="hint r" style="--i:4">press <kbd>→</kbd> to start</p>
+</div>
