+++ mono/src/lib/slides/SlideJourney.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+  import Pipeline from "./Pipeline.svelte";
+</script>
+
+<div class="slide-body wide">
+  <span class="eyebrow r" style="--i:0">The path</span>
+  <h2 class="title r" style="--i:1">From raw data to delivered exam</h2>
+  <p class="message r" style="--i:2">The modules chain into a natural pipeline.</p>
+  <Pipeline />
+  <p class="aside r" style="--i:6">
+    Each step hands clean data to the next — move forward at your own pace.
+  </p>
+</div>
